Add .nojekyll file to bypass Jekyll build
This commit adds a .nojekyll file to the root of the repository. By including this file, we disable the default Jekyll processing on GitHub Pages. 

Since our site does not use Jekyll and relies solely on static HTML (index.html), this prevents GitHub Pages from attempting to run Jekyll builds, which could cause deployment issues. Adding this file ensures that the content is deployed directly as intended.